refactor: resolve #141 - split remaining files over 500-line limit
Extract focused modules from 3 files exceeding the 500-line limit:
- useBasicIdeMessageHandlers.ts (674→379): extract screenHandlers.ts + messageHandlerContext.ts
- MusicDSLParser.ts (593→399): extract musicValidation.ts
- AnimationWorker.ts (568→286): extract animationSyncHandlers.ts

Closes #141
