+++ src/features/ide/composables/useBasicIdeMessageHandlers.ts
@@ -3,40 +3,24 @@
  *
  * Use loglevel: log.getLogger('ide-messages').setLevel('debug') in console for verbose logs.
  */
-import type { Ref } from 'vue'
-
-import type { SharedDisplayViews } from '@/core/animation/sharedDisplayBuffer'
-import type { DecodedScreenState } from '@/core/animation/sharedDisplayBufferAccessor'
-import type {
-  AnyServiceWorkerMessage,
-  ErrorMessage,
-  RequestInputMessage,
-  ResultMessage,
-  ScreenCell,
-} from '@/core/interfaces'
+import type { AnyServiceWorkerMessage, ErrorMessage, ResultMessage } from '@/core/interfaces'
 import type { Note, Rest } from '@/core/sound/types'
-import type { SpriteState } from '@/core/sprite/types'
-import { clearBackgroundTileCache } from '@/features/ide/composables/useCanvasBackgroundRenderer'
-import { clearSpriteImageCache } from '@/features/ide/composables/useKonvaSpriteRenderer'
 import { ExecutionError } from '@/features/ide/errors/ExecutionError'
-import { setRuntimePaletteCombination } from '@/shared/data/palette'
 import { logComposable, logIdeMessages } from '@/shared/logger'
 
-import type { WebWorkerManager } from './useBasicIdeWebWorkerUtils'
+import type { MessageHandlerContext,QueuedMessage } from './messageHandlerContext'
+import { handleScreenChangedMessage, handleScreenUpdateMessage, handleSpriteStatesMessage } from './screenHandlers'
 import { extendExecutionTimeout } from './useBasicIdeWebWorkerUtils'
 import { useWebAudioPlayer } from './useWebAudioPlayer'
 
 export { ExecutionError }
+export type { MessageHandlerContext, PendingSpriteAction, SpriteActionQueues } from './messageHandlerContext'
+export { handleScreenUpdateMessage } from './screenHandlers'
 
 /**
  * Message queue for non-critical messages
  * Processed during requestAnimationFrame to align with rendering
  */
-interface QueuedMessage {
-  message: AnyServiceWorkerMessage
-  context: MessageHandlerContext
-}
-
 const messageQueue: QueuedMessage[] = []
 let isProcessingQueue = false
 let queueAnimationFrame: number | null = null
@@ -135,52 +119,6 @@ function flushMessageQueue(): void {
 }
 
 /**
- * Pending action for a sprite when its Konva node does not exist yet;
- * applied when node is created or START_MOVEMENT is handled.
- */
-export type PendingSpriteAction = { type: 'POSITION'; x: number; y: number }
-
-/** Per-sprite action queue (action number → list of pending actions). */
-export type SpriteActionQueues = Map<number, PendingSpriteAction[]>
-
-export interface MessageHandlerContext {
-  output: Ref<string[]>
-  errors: Ref<
-    Array<{ line: number; message: string; type: string; stack?: string; sourceLine?: string }>
-  >
-  debugOutput: Ref<string>
-  screenBuffer: Ref<ScreenCell[][]>
-  cursorX: Ref<number>
-  cursorY: Ref<number>
-  bgPalette: Ref<number>
-  backdropColor?: Ref<number>
-  spritePalette?: Ref<number>
-  cgenMode?: Ref<number>
-  /** Sprite states from DEF SPRITE and SPRITE commands (updated via SPRITE_STATES message) */
-  spriteStates?: Ref<SpriteState[]>
-  /** Whether sprite display is enabled (SPRITE ON/OFF) */
-  spriteEnabled?: Ref<boolean>
-  // movementStates removed - read from shared buffer instead
-  frontSpriteNodes?: Ref<Map<number, unknown>>
-  backSpriteNodes?: Ref<Map<number, unknown>>
-  /** Per-sprite action queue; POSITION etc. when node does not exist; consumed when START_MOVEMENT is handled. */
-  spriteActionQueues?: Ref<SpriteActionQueues>
-  webWorkerManager: WebWorkerManager
-  /** Shared display buffer views; main reads screen/cursor/scalars when SCREEN_CHANGED. */
-  sharedDisplayViews?: SharedDisplayViews
-  /** Called when SCREEN_CHANGED is received to schedule a render (Screen.vue reads from shared buffer). */
-  scheduleRender?: () => void
-  /** Coalesced version: at most one schedule per frame. Prefer this for SCREEN_CHANGED to avoid main-thread flood. */
-  scheduleRenderForScreenChanged?: () => void
-  /** Called by Screen.vue after decoding shared buffer to update refs (screenBuffer, cursorX, etc.). */
-  setDecodedScreenState?: (decoded: DecodedScreenState) => void
-  /** Pending INPUT/LINPUT request from worker; set when REQUEST_INPUT is received, cleared on submit/cancel. */
-  pendingInputRequest?: Ref<RequestInputMessage['data'] | null>
-  /** Send INPUT_VALUE to worker to resolve a pending input request. */
-  respondToInputRequest?: (requestId: string, values: string[], cancelled: boolean) => void
-}
-
-/**
  * Handle output message from web worker
  */
 export function handleOutputMessage(message: AnyServiceWorkerMessage, context: MessageHandlerContext): void {
@@ -198,212 +136,6 @@ export function handleOutputMessage(message: AnyServiceWorkerMessage, context: M
       message: outputText,
       type: 'runtime',
     })
-  }
-}
-
-/**
- * Handle SCREEN_CHANGED message from web worker (shared buffer path).
- * Uses coalesced schedule so many SCREEN_CHANGED (e.g. loop with PRINT) only trigger one render per frame.
- */
-export function handleScreenChangedMessage(
-  _message: AnyServiceWorkerMessage,
-  context: MessageHandlerContext
-): void {
-  const schedule = context.scheduleRenderForScreenChanged ?? context.scheduleRender
-  if (schedule) {
-    schedule()
-  }
-}
-
-/**
- * Handle SPRITE_STATES message from web worker.
- * Updates sprite states and triggers a render to display static sprites (DEF SPRITE, SPRITE commands).
- */
-export function handleSpriteStatesMessage(
-  message: AnyServiceWorkerMessage,
-  context: MessageHandlerContext
-): void {
-  if (message.type !== 'SPRITE_STATES') return
-
-  const { spriteStates, spriteEnabled } = message.data
-  logIdeMessages.debug('📤 Handling SPRITE_STATES:', { statesCount: spriteStates?.length, enabled: spriteEnabled })
-
-  // Update sprite states in context (these are passed to Screen.vue via useScreenContext)
-  if (context.spriteStates && spriteStates) {
-    context.spriteStates.value = spriteStates
-  }
-  if (context.spriteEnabled !== undefined && spriteEnabled !== undefined) {
-    context.spriteEnabled.value = spriteEnabled
-  }
-
-  // Schedule a full render to display the sprites
-  const schedule = context.scheduleRender
-  if (schedule) {
-    schedule()
-  }
-}
-
-/**
- * Handle screen update message from web worker (legacy SCREEN_UPDATE; worker no longer sends for shared buffer path)
- */
-export function handleScreenUpdateMessage(message: AnyServiceWorkerMessage, context: MessageHandlerContext): void {
-  if (message.type !== 'SCREEN_UPDATE') return
-
-  const update = message.data
-  if (!update) {
-    logComposable.warn('SCREEN_UPDATE message has no data')
-    return
-  }
-  if (!context?.screenBuffer) {
-    logComposable.warn('SCREEN_UPDATE: context or screenBuffer missing, skipping')
-    return
-  }
-
-  switch (update.updateType) {
-    case 'character':
-      if (update.x !== undefined && update.y !== undefined && update.character !== undefined) {
-        const x = update.x
-        const y = update.y
-        const character = update.character
-
-        // Ensure row exists
-        context.screenBuffer.value[y] ??= []
-
-        // Ensure cell exists
-        const currentRow = context.screenBuffer.value[y]
-        currentRow[x] ??= {
-          character: ' ',
-          colorPattern: 0,
-          x,
-          y,
-        }
-
-        // Update character - force reactivity by creating new object
-        const currentCell = currentRow[x]
-        const newCell: ScreenCell = {
-          character,
-          colorPattern: currentCell.colorPattern,
-          x: currentCell.x,
-          y: currentCell.y,
-        }
-        currentRow[x] = newCell
-
-        // Also trigger reactivity by reassigning the row
-        context.screenBuffer.value[y] = [...currentRow]
-      }
-      break
-    case 'cursor':
-      if (update.cursorX !== undefined) context.cursorX.value = update.cursorX
-      if (update.cursorY !== undefined) context.cursorY.value = update.cursorY
-      break
-    case 'clear':
-      // Clear screen buffer
-      for (let y = 0; y < 24; y++) {
-        const row = context.screenBuffer.value[y]
-        if (row) {
-          for (let x = 0; x < 28; x++) {
-            const cell = row[x]
-            if (cell) {
-              cell.character = ' '
-            }
-          }
-        }
-      }
-      context.cursorX.value = 0
-      context.cursorY.value = 0
-      break
-    case 'full':
-      if (update.screenBuffer) {
-        context.screenBuffer.value = update.screenBuffer
-      }
-      if (update.cursorX !== undefined) context.cursorX.value = update.cursorX
-      if (update.cursorY !== undefined) context.cursorY.value = update.cursorY
-      break
-    case 'color':
-      // Update color pattern for cells specified in colorUpdates
-      if (update.colorUpdates) {
-        for (const colorUpdate of update.colorUpdates) {
-          const { x, y, pattern } = colorUpdate
-
-          // Ensure row exists
-          context.screenBuffer.value[y] ??= []
-
-          // Ensure cell exists
-          const currentRow = context.screenBuffer.value[y]
-          currentRow[x] ??= {
-            character: ' ',
-            colorPattern: 0,
-            x,
-            y,
-          }
-
-          // Update color pattern - force reactivity by creating new object
-          const currentCell = currentRow[x]
-          const newCell: ScreenCell = {
-            character: currentCell.character,
-            colorPattern: pattern,
-            x: currentCell.x,
-            y: currentCell.y,
-          }
-          currentRow[x] = newCell
-
-          // Also trigger reactivity by reassigning the row
-          context.screenBuffer.value[y] = [...currentRow]
-        }
-      }
-      break
-    case 'palette':
-      // Update background and sprite palette codes
-      if (update.bgPalette !== undefined) {
-        context.bgPalette.value = update.bgPalette
-        logComposable.debug('Updated background palette:', update.bgPalette)
-      }
-      if (update.spritePalette !== undefined) {
-        // Note: spritePalette is stored but not currently used in rendering
-        // It will be used when sprite system is implemented
-        logComposable.debug('Updated sprite palette:', update.spritePalette)
-      }
-      break
-    case 'backdrop':
-      // Update backdrop color
-      if (update.backdropColor !== undefined && context.backdropColor) {
-        context.backdropColor.value = update.backdropColor
-        logComposable.debug('Updated backdrop color:', update.backdropColor)
-      }
-      break
-    case 'cgen':
-      // Update character generator mode
-      if (update.cgenMode !== undefined) {
-        if (context.cgenMode) {
-          context.cgenMode.value = update.cgenMode
-        }
-        logComposable.debug('Updated character generator mode:', update.cgenMode)
-      }
-      break
-    case 'palette-combination':
-      if (
-        update.paletteTarget &&
-        update.paletteIndex !== undefined &&
-        update.paletteCombination !== undefined &&
-        update.paletteColors
-      ) {
-        setRuntimePaletteCombination(
-          update.paletteTarget,
-          update.paletteIndex,
-          update.paletteCombination,
-          update.paletteColors
-        )
-        clearBackgroundTileCache()
-        clearSpriteImageCache()
-        context.scheduleRender?.()
-        logComposable.debug('Updated runtime palette combination:', {
-          target: update.paletteTarget,
-          paletteIndex: update.paletteIndex,
-          combination: update.paletteCombination,
-          colors: update.paletteColors,
-        })
-      }
-      break
   }
 }
 
@@ -551,33 +283,6 @@ export function handlePlaySoundMessage(message: AnyServiceWorkerMessage, _contex
 
   // Play sequentially: next PLAY starts after current melody finishes (F-BASIC behavior)
   audioPlayer.playMusicSequential(channels)
-}
-
-/**
- * Handle animation command message from web worker
- *
- * NOTE: All animation commands are now handled via direct sync (Executor Worker → Animation Worker)
- * through shared buffer using Atomics. Main thread reads animation state ONLY from shared buffer.
- *
- * Per real F-BASIC hardware behavior:
- * - POSITION on inactive sprite: No immediate screen change; stores as start position for next MOVE
- * - POSITION on active sprite: Takes effect immediately (render loop reads from shared buffer)
- *
- * This handler is obsolete and should no longer receive messages.
- */
-// Omitted: handleAnimationCommandMessage - Animation commands now handled via shared buffer sync
-
-/**
-    case 2:
-    case 8:
-      return -1 // Up directions
-    case 4:
-    case 5:
-    case 6:
-      return 1 // Down directions
-    default:
-      return 0
-  }
 }
 
 /**
